componetizded the mobile navbar. Will look into reducing the code more
