--- client/src/pages/Intro/Nav/index.js
@@ -1,2 +1,0 @@
-import Nav from './Nav';
-export default Nav;
