+++ client/src/pages/Report/SleeperWeeklyReport.js
@@ -25,7 +25,7 @@ const SleeperWeeklyReport = (props) => {
     <>
       <div style={{ textAlign: "center" }}>
         <button className="btn btn--sleeper u-margin-top" onClick={() => exportComponentAsPNG(componentRef)}>
-          Click here for your report to be converted to an image!
+          Click here to download your report as an image!
         </button>
       </div>
