+++ client/src/sass/pages/intro/_composition.scss
@@ -35,7 +35,6 @@
         }
 
         &:hover {
-            outline: 1.5rem solid $primary-color;
             transform: scale(1.05) translateY(5px);
             box-shadow: 0 2.5rem 4 rem rgba($black-color, .5);
             z-index: 2;
